backend/requirements.txt
# FastAPI Backend Dependencies (Python 3.13 compatible)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.10
python-multipart>=0.0.17
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.28.0
pandas>=2.2.3
alembic>=1.14.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
